[expo-updates][ios][1/n] Refactor responsibility
